refactor: split chat.rs into chat/ module (SRP)
