+++ src/ui/chat/mod.rs
@@ -1,0 +1,6 @@
+mod actions;
+mod menu;
+mod models;
+mod view;
+
+pub use view::ChatView;
